+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added custom color support
 * Added space support
+* Added custom space support
 
 ## 0.0.5 - 2024-10-16
 
@@ -9,8 +10,7 @@
 
 ## 0.0.4 - 2024-10-16
 
-* Added spacing/insets utility
-* Added inline documentation
+* Added insets support (padding/margin)
 
 ## 0.0.1 - 2024-10-16